refactor: dedupe formatEvaluationDate into a shared util
Decisions and DecisionInstance each defined an identical
formatEvaluationDate — moved to operate/shared/utils/ and imported by
both instead.
