--- webapp/client/apps/orchestration-cluster-webapp/src/operate/pages/DecisionInstance/utils.ts
@@ -1,15 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-
-import {format, parseISO} from 'date-fns';
-
-function formatEvaluationDate(dateString: string | null | undefined): string {
-	return dateString ? format(parseISO(dateString), 'yyyy-MM-dd HH:mm:ss') : '--';
-}
-
-export {formatEvaluationDate};
